Fixes #35684 - Drop Applied catalog lines
In the past we dropped the line which stated how log catalog application
took since there's a metric for this. This was done by matching the
content, but the content changed somewhere between report version 3 and
6.

This saves a lot of redundant messages in the Foreman database. An
uneventful report will now have no messages at all.