Rewrite ENC to use classes
This rewrite uses classes to coherently group functionality. It attempts
to separate out the actual client logic into a ForemanClient class. Then
around that it wraps methods to push and watch as well as the ENC.

Uploading is rewritten to always use the same logic, regardless of
whether it's single threaded or parallel. Single threaded just spawns a
single thread. This also has the advantage of being able to reuse HTTP
connection instances, possibly leading to higher throughput since it
doesn't have to establish fresh HTTP(S) connections for every request.