puppet/systemd: Allow 7.x
In https://github.com/theforeman/puppet-puppetserver_foreman/pull/52 we
added a dependency to puppet/systemd. After some digging around we
identified https://github.com/voxpupuli/puppet-systemd/pull/433 which is
probably the newest requirement we have. That means we don't need to
depend on current latest (8.2.0) version of the module, 7.0.0 and newer
are fine.